git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/MPIDataMining@2186 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/tools/GridPrinterForStretching.cpp
@@ -29,11 +29,11 @@ GridPrinterForStretching::~GridPrinterForStretching()
 {
 }
 
-void GridPrinterForStretching::printGridDomain(DataVector& alpha, std::string tFilename, BoundingBox& GridArea, double PointsPerDimension){
+void GridPrinterForStretching::printGridDomain(DataVector& alpha, std::string tFilename, BoundingBox& GridArea, size_t PointsPerDimension){
 	throw new tool_exception("GridPrinterForStretching::printGridDomain : This printer does not support BoundingBox, use printGridDomainStretching instead!");
 }
 
-void GridPrinterForStretching::printGridDomainStretching(DataVector& alpha, std::string tFilename, Stretching& GridArea, double PointsPerDimension)
+void GridPrinterForStretching::printGridDomainStretching(DataVector& alpha, std::string tFilename, Stretching& GridArea, size_t PointsPerDimension)
 {
 	DimensionBoundary dimOne;
 	DimensionBoundary dimTwo;
@@ -54,9 +54,9 @@ void GridPrinterForStretching::printGridDomainStretching(DataVector& alpha, std:
 			dimOne = GridArea.getBoundary(0);
 			dimTwo = GridArea.getBoundary(1);
 
-			for (double i = dimOne.leftBoundary; i <= dimOne.rightBoundary; i+=((dimOne.rightBoundary - dimOne.leftBoundary)/PointsPerDimension))
+			for (double i = dimOne.leftBoundary; i <= dimOne.rightBoundary; i+=((dimOne.rightBoundary - dimOne.leftBoundary)/static_cast<double>(PointsPerDimension)))
 			{
-				for (double j = dimTwo.leftBoundary; j <= dimTwo.rightBoundary; j+=((dimTwo.rightBoundary - dimTwo.leftBoundary)/PointsPerDimension))
+				for (double j = dimTwo.leftBoundary; j <= dimTwo.rightBoundary; j+=((dimTwo.rightBoundary - dimTwo.leftBoundary)/static_cast<double>(PointsPerDimension)))
 				{
 					std::vector<double> point;
 					point.push_back(i);
@@ -77,7 +77,7 @@ void GridPrinterForStretching::printGridDomainStretching(DataVector& alpha, std:
 	}
 }
 
-void GridPrinterForStretching::printGrid(DataVector& alpha, std::string tFilename, double PointsPerDimension)
+void GridPrinterForStretching::printGrid(DataVector& alpha, std::string tFilename, size_t PointsPerDimension)
 {
 		DimensionBoundary dimOne;
 		DimensionBoundary dimTwo;
@@ -99,9 +99,9 @@ void GridPrinterForStretching::printGrid(DataVector& alpha, std::string tFilenam
 				{
 					dimOne = myGrid->getStretching()->getBoundary(0);
 
-					double inc_x = ((dimOne.rightBoundary - dimOne.leftBoundary)/(PointsPerDimension-1));
+					double inc_x = ((dimOne.rightBoundary - dimOne.leftBoundary)/(static_cast<double>(PointsPerDimension)-1.0));
 
-					size_t points = (size_t)PointsPerDimension;
+					size_t points = PointsPerDimension;
 
 					for (size_t i = 0; i < points; i++)
 					{
@@ -117,10 +117,10 @@ void GridPrinterForStretching::printGrid(DataVector& alpha, std::string tFilenam
 
 					double offset_x = dimOne.leftBoundary;
 					double offset_y = dimTwo.leftBoundary;
-					double inc_x = ((dimOne.rightBoundary - dimOne.leftBoundary)/(PointsPerDimension-1));
-					double inc_y = ((dimTwo.rightBoundary - dimTwo.leftBoundary)/(PointsPerDimension-1));
+					double inc_x = ((dimOne.rightBoundary - dimOne.leftBoundary)/(static_cast<double>(PointsPerDimension)-1.0));
+					double inc_y = ((dimTwo.rightBoundary - dimTwo.leftBoundary)/(static_cast<double>(PointsPerDimension)-1.0));
 
-					size_t points = (size_t)PointsPerDimension;
+					size_t points = PointsPerDimension;
 
 					for (size_t i = 0; i < points; i++)
 					{
